api/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
psycopg[binary]==3.2.4
PyJWT==2.9.0
httpx==0.27.2
python-multipart==0.0.9
